fix: reactflow folder moved to components
