+++ src/components/reactflow/reactflow.config.tsx
@@ -11,8 +11,8 @@ import {
 } from '@xyflow/react'
 import '@xyflow/react/dist/style.css'
 
-import FloatingEdge from '@/app/(frontend)/(dashboard)/reactflow/FloatingEdges'
-import FloatingConnectionLine from '@/app/(frontend)/(dashboard)/reactflow/FloatingEdges/FloatingConnectionLine'
+import FloatingEdge from '@/components/reactflow/FloatingEdges'
+import FloatingConnectionLine from '@/components/reactflow/FloatingEdges/FloatingConnectionLine'
 import { cn } from '@/lib/utils'
 
 import CustomNode from './CustomNodes'
